Add booking and request, refactor
