+++ src/main/java/ru/practicum/shareit/booking/interfaces/BookingStorage.java
@@ -1,0 +1,21 @@
+package ru.practicum.shareit.booking.interfaces;
+
+import java.util.List;
+import java.util.Optional;
+
+import ru.practicum.shareit.booking.dto.BookingDto;
+import ru.practicum.shareit.booking.model.Booking;
+
+public interface BookingStorage {
+    Optional<BookingDto> add(Booking booking);
+
+    Optional<BookingDto> get(Long id);
+
+    List<BookingDto> getByUserId(Long userId);
+
+    Optional<BookingDto> update(Booking booking);
+
+    void delete(Long id);
+
+    boolean checkIdExist(Long id);
+}
