+++ src/main/java/ru/practicum/shareit/booking/model/Booking.java
@@ -1,0 +1,19 @@
+package ru.practicum.shareit.booking.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import ru.practicum.shareit.item.model.Item;
+import ru.practicum.shareit.user.model.User;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Booking {
+    private Long id;
+    private java.time.LocalDateTime start;
+    private java.time.LocalDateTime end;
+    private Item item;
+    private User booker;
+    private BookingStatus status;
+}
