--- src/main/java/ru/practicum/shareit/exception/AccessDeniedException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class AccessDeniedException extends RuntimeException {
-    public AccessDeniedException(String message) {
-        super(message);
-    }
-}
