--- src/main/java/ru/practicum/shareit/exception/ConditionsNotMetException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class ConditionsNotMetException extends RuntimeException {
-    public ConditionsNotMetException(String message) {
-        super(message);
-    }
-}
