--- src/main/java/ru/practicum/shareit/exception/ConstraintViolationException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class ConstraintViolationException extends RuntimeException {
-    public ConstraintViolationException(String message) {
-        super(message);
-    }
-}
