--- src/main/java/ru/practicum/shareit/exception/DuplicatedDataException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class DuplicatedDataException extends RuntimeException {
-    public DuplicatedDataException(String message) {
-        super(message);
-    }
-}
