--- src/main/java/ru/practicum/shareit/exception/ErrorResponse.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    public String error;
-}
