--- src/main/java/ru/practicum/shareit/exception/HttpMethodNotSupportedException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class HttpMethodNotSupportedException extends RuntimeException {
-    public HttpMethodNotSupportedException(String message) {
-        super(message);
-    }
-}
