--- src/main/java/ru/practicum/shareit/exception/InternalServerException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class InternalServerException extends RuntimeException {
-    public InternalServerException(String message) {
-        super(message);
-    }
-}
