--- src/main/java/ru/practicum/shareit/exception/MethodArgumentNotValidException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class MethodArgumentNotValidException extends RuntimeException {
-    public MethodArgumentNotValidException(String message) {
-        super(message);
-    }
-}
