--- src/main/java/ru/practicum/shareit/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
