+++ src/main/java/ru/practicum/shareit/request/controller/ItemRequestController.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.request.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * TODO Sprint add-item-requests.
+ */
+@RestController
+@RequestMapping(path = "/requests")
+public class ItemRequestController {
+}
