+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,0 +1,17 @@
+package ru.practicum.shareit.request.model;
+
+import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ItemRequest {
+    private Long id;
+    private String description;
+    private Long requestor;
+    private LocalDateTime created;
+}
