+++ src/main/java/ru/practicum/shareit/system/ShareItApp.java
@@ -1,0 +1,13 @@
+package ru.practicum.shareit.system;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ShareItApp {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ShareItApp.class, args);
+	}
+
+}
