+++ src/main/java/ru/practicum/shareit/system/exception/AccessDeniedException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class AccessDeniedException extends RuntimeException {
+    public AccessDeniedException(String message) {
+        super(message);
+    }
+}
