+++ src/main/java/ru/practicum/shareit/system/exception/ConditionsNotMetException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class ConditionsNotMetException extends RuntimeException {
+    public ConditionsNotMetException(String message) {
+        super(message);
+    }
+}
