+++ src/main/java/ru/practicum/shareit/system/exception/ConstraintViolationException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class ConstraintViolationException extends RuntimeException {
+    public ConstraintViolationException(String message) {
+        super(message);
+    }
+}
