+++ src/main/java/ru/practicum/shareit/system/exception/DuplicatedDataException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class DuplicatedDataException extends RuntimeException {
+    public DuplicatedDataException(String message) {
+        super(message);
+    }
+}
