+++ src/main/java/ru/practicum/shareit/system/exception/ErrorResponse.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.system.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+    public String error;
+}
