+++ src/main/java/ru/practicum/shareit/system/exception/HttpMethodNotSupportedException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class HttpMethodNotSupportedException extends RuntimeException {
+    public HttpMethodNotSupportedException(String message) {
+        super(message);
+    }
+}
