+++ src/main/java/ru/practicum/shareit/system/exception/InternalServerException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class InternalServerException extends RuntimeException {
+    public InternalServerException(String message) {
+        super(message);
+    }
+}
