+++ src/main/java/ru/practicum/shareit/system/exception/MethodArgumentNotValidException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class MethodArgumentNotValidException extends RuntimeException {
+    public MethodArgumentNotValidException(String message) {
+        super(message);
+    }
+}
