+++ src/main/java/ru/practicum/shareit/system/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
