+++ src/main/java/ru/practicum/shareit/system/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.system.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String message) {
+        super(message);
+    }
+}
