--- src/main/java/ru/practicum/shareit/user/UniqueConstraint.java
@@ -1,5 +1,0 @@
-package ru.practicum.shareit.user;
-
-public @interface UniqueConstraint {
-
-}
