--- src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,51 +1,0 @@
-package ru.practicum.shareit.user;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-import ru.practicum.shareit.user.interfaces.UserService;
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.utils.Validate;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping(path = "/users")
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public UserDto get(@PathVariable Long id) {
-        return userService.get(id);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.OK)
-    public UserDto add(@RequestBody UserDto user) {
-        Validate.user(user);
-        return userService.add(user);
-    }
-
-    @PatchMapping("/{userId}")
-    @ResponseStatus(HttpStatus.OK)
-    public UserDto patch(@PathVariable Long userId, @RequestBody UserDto user) {
-        user.setId(userId);
-        Validate.user(user);
-        return userService.patch(user);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void delete(@PathVariable Long id) {
-        userService.delete(id);
-    }
-}
