+++ src/main/java/ru/practicum/shareit/user/interfaces/UniqueConstraint.java
@@ -1,0 +1,5 @@
+package ru.practicum.shareit.user.interfaces;
+
+public @interface UniqueConstraint {
+
+}
