+++ src/main/java/ru/practicum/shareit/user/mapper/UserMapper.java
@@ -8,25 +8,25 @@ import ru.practicum.shareit.user.model.User;
 
 @UtilityClass
 public class UserMapper {
-    public static UserDto userToDto(User user) {
+    public static UserDto toDto(User user) {
         return new UserDto(
                 user.getId(),
                 user.getName(),
                 user.getEmail());
     }
 
-    public static User dtoToUser(UserDto userDto) {
+    public static User toUser(UserDto userDto) {
         return new User(
                 userDto.getId(),
                 userDto.getName(),
                 userDto.getEmail());
     }
 
-    public static List<User> dtoToUser(List<UserDto> usersDto) {
-        return usersDto.stream().map(UserMapper::dtoToUser).toList();
+    public static List<User> toUser(List<UserDto> usersDto) {
+        return usersDto.stream().map(UserMapper::toUser).toList();
     }
 
-    public static List<UserDto> userToDto(List<User> users) {
-        return users.stream().map(UserMapper::userToDto).toList();
+    public static List<UserDto> toDto(List<User> users) {
+        return users.stream().map(UserMapper::toDto).toList();
     }
 }
